+++ src/template/ompcompress.c
@@ -39,9 +39,9 @@ _t2(compress_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
       p += x;
       /* compress partial or full block */
       if (nx - x < 4u)
-        block_size = _t2(zfp_encode_partial_block_strided, Scalar, 1)(&s, p, nx - x, 1);
+        block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 1)(&s, p, nx - x, 1);
       else
-        block_size = _t2(zfp_encode_block, Scalar, 1)(&s, p);
+        block_size = (uint16)_t2(zfp_encode_block, Scalar, 1)(&s, p);
       if (length_table)
         length_table[block] = block_size;
     }
@@ -91,9 +91,9 @@ _t2(compress_strided_omp, Scalar, 1)(zfp_stream* stream, const zfp_field* field)
       p += sx * (ptrdiff_t)x;
       /* compress partial or full block */
       if (nx - x < 4u)
-        block_size = _t2(zfp_encode_partial_block_strided, Scalar, 1)(&s, p, nx - x, sx);
+        block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 1)(&s, p, nx - x, sx);
       else
-        block_size = _t2(zfp_encode_block_strided, Scalar, 1)(&s, p, sx);
+        block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 1)(&s, p, sx);
       if (length_table)
         length_table[block] = block_size;
     }
@@ -150,9 +150,9 @@ _t2(compress_strided_omp, Scalar, 2)(zfp_stream* stream, const zfp_field* field)
       p += sx * (ptrdiff_t)x + sy * (ptrdiff_t)y;
       /* compress partial or full block */
       if (nx - x < 4u || ny - y < 4u)
-        block_size = _t2(zfp_encode_partial_block_strided, Scalar, 2)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), sx, sy);
+        block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 2)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), sx, sy);
       else
-        block_size = _t2(zfp_encode_block_strided, Scalar, 2)(&s, p, sx, sy);
+        block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 2)(&s, p, sx, sy);
       if (length_table)
         length_table[block] = block_size;
     }
@@ -213,9 +213,9 @@ _t2(compress_strided_omp, Scalar, 3)(zfp_stream* stream, const zfp_field* field)
       p += sx * (ptrdiff_t)x + sy * (ptrdiff_t)y + sz * (ptrdiff_t)z;
       /* compress partial or full block */
       if (nx - x < 4u || ny - y < 4u || nz - z < 4u)
-        block_size = _t2(zfp_encode_partial_block_strided, Scalar, 3)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), sx, sy, sz);
+        block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 3)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), sx, sy, sz);
       else
-        block_size = _t2(zfp_encode_block_strided, Scalar, 3)(&s, p, sx, sy, sz);
+        block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 3)(&s, p, sx, sy, sz);
       if (length_table)
         length_table[block] = block_size;
     }
@@ -280,9 +280,9 @@ _t2(compress_strided_omp, Scalar, 4)(zfp_stream* stream, const zfp_field* field)
       p += sx * (ptrdiff_t)x + sy * (ptrdiff_t)y + sz * (ptrdiff_t)z + sw * (ptrdiff_t)w;
       /* compress partial or full block */
       if (nx - x < 4u || ny - y < 4u || nz - z < 4u || nw - w < 4u)
-        block_size = _t2(zfp_encode_partial_block_strided, Scalar, 4)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), MIN(nw - w, 4u), sx, sy, sz, sw);
+        block_size = (uint16)_t2(zfp_encode_partial_block_strided, Scalar, 4)(&s, p, MIN(nx - x, 4u), MIN(ny - y, 4u), MIN(nz - z, 4u), MIN(nw - w, 4u), sx, sy, sz, sw);
       else
-        block_size = _t2(zfp_encode_block_strided, Scalar, 4)(&s, p, sx, sy, sz, sw);
+        block_size = (uint16)_t2(zfp_encode_block_strided, Scalar, 4)(&s, p, sx, sy, sz, sw);
       if (length_table)
         length_table[block] = block_size;
     }
